Move enums to base package
