+++ src/main/java/net/tridentsdk/entity/meta/living/VillagerMeta.java
@@ -1,9 +1,6 @@
 package net.tridentsdk.entity.meta.living;
 
-import lombok.Getter;
-
-import java.util.HashMap;
-import java.util.Map;
+import net.tridentsdk.base.VillagerProfession;
 
 /**
  * @author TridentSDK
@@ -14,38 +11,5 @@ public interface VillagerMeta extends AgeableMeta{
     VillagerProfession getProfession();
 
     void setProfession(VillagerProfession profession);
-
-    enum VillagerProfession {
-
-        FARMER(0),
-        LIBRARIAN(1),
-        PRIEST(2),
-        BLACKSMITH(3),
-        BUTCHER(4);
-
-        @Getter
-        private final int data;
-
-        private VillagerProfession(int data) {
-            this.data = data;
-        }
-
-        private static final Map<Integer, VillagerProfession> dataToType = new HashMap<>();
-
-        public static VillagerProfession of(int data) {
-            VillagerProfession profession = dataToType.get(data);
-            if (profession == null) {
-                throw new IllegalArgumentException("no villager profession with id = " + data);
-            }
-            return profession;
-        }
-
-        static {
-            for (VillagerProfession type : values()) {
-                dataToType.put(type.data, type);
-            }
-        }
-
-    }
 
 }
