+++ src/main/java/net/tridentsdk/entity/meta/living/animal/OcelotMeta.java
@@ -1,5 +1,7 @@
 package net.tridentsdk.entity.meta.living.animal;
 
+import net.tridentsdk.base.OcelotType;
+
 /**
  * @author TridentSDK
  * @since 0.5-alpha
@@ -9,9 +11,5 @@ public interface OcelotMeta extends TameableAnimalMeta {
     OcelotType getOcelotType();
 
     void setOcelotType(OcelotType type);
-
-    enum OcelotType {
-        // TODO
-    }
 
 }
