+++ src/main/java/net/tridentsdk/entity/meta/living/animal/RabbitMeta.java
@@ -1,5 +1,7 @@
 package net.tridentsdk.entity.meta.living.animal;
 
+import net.tridentsdk.base.RabbitType;
+
 /**
  * @author TridentSDK
  * @since 0.5-alpha
@@ -9,9 +11,5 @@ public interface RabbitMeta extends AnimalMeta {
     RabbitType getRabbitType();
 
     void setRabbitType(RabbitType type);
-
-    enum RabbitType {
-        // TODO
-    }
 
 }
