+++ src/main/java/net/tridentsdk/entity/meta/living/monster/SkeletonMeta.java
@@ -1,9 +1,6 @@
 package net.tridentsdk.entity.meta.living.monster;
 
-import lombok.Getter;
-
-import java.util.HashMap;
-import java.util.Map;
+import net.tridentsdk.base.SkeletonType;
 
 /**
  * @author TridentSDK
@@ -18,36 +15,5 @@ public interface SkeletonMeta extends MonsterMeta {
     boolean isSwingingArms();
 
     void setSwingingArms(boolean swingingArms);
-
-    enum SkeletonType {
-
-        NORMAL(0),
-        WITHER(1),
-        STRAY(2);
-
-        @Getter
-        private final int data;
-
-        private SkeletonType(int data) {
-            this.data = data;
-        }
-
-        private static final Map<Integer, SkeletonType> dataToType = new HashMap<>();
-
-        public static SkeletonType of(int data) {
-            SkeletonType type = dataToType.get(data);
-            if (type == null) {
-                throw new IllegalArgumentException("no skeleton type with id = " + data);
-            }
-            return type;
-        }
-
-        static {
-            for (SkeletonType type : values()) {
-                dataToType.put(type.data, type);
-            }
-        }
-
-    }
 
 }
