+++ src/main/java/net/tridentsdk/entity/meta/living/monster/ZombieMeta.java
@@ -1,9 +1,6 @@
 package net.tridentsdk.entity.meta.living.monster;
 
-import lombok.Getter;
-
-import java.util.HashMap;
-import java.util.Map;
+import net.tridentsdk.base.ZombieType;
 
 /**
  * @author TridentSDK
@@ -26,44 +23,5 @@ public interface ZombieMeta extends MonsterMeta {
     boolean areHandsHeldUp();
 
     void setHandsHeldUp(boolean handsUp);
-
-    enum ZombieType {
-
-        NORMAL(0),
-        VILLAGER_FARMER(1),
-        VILLAGER_LIBRARIAN(2),
-        VILLAGER_PRIEST(3),
-        VILLAGER_BLACKSMITH(4),
-        VILLAGER_BUTCHER(5),
-        HUSK(6);
-
-        @Getter
-        private final int data;
-
-        @Getter
-        private final boolean villager;
-
-        private ZombieType(int data) {
-            this.data = data;
-            this.villager = data >= 1 && data <= 5;
-        }
-
-        private static final Map<Integer, ZombieType> dataToType = new HashMap<>();
-
-        public static ZombieType of(int data) {
-            ZombieType type = dataToType.get(data);
-            if (type == null) {
-                throw new IllegalArgumentException("no zombie type with id = " + data);
-            }
-            return type;
-        }
-
-        static {
-            for (ZombieType type : values()) {
-                dataToType.put(type.data, type);
-            }
-        }
-
-    }
 
 }
